TestGit: extract interface GrowingGraphExample
